+++ packages/js/src/index.js
@@ -1,5 +1,9 @@
 import {CoreManager, Constants} from '@skywatch/api';
+import {CameraViewWebComponent} from '@skywatch/react';
 const {SERVER_URL, API_KEY} = Constants;
+
+customElements.get('camera-view-web-component') ||
+  customElements.define('camera-view-web-component', CameraViewWebComponent);
 
 const Skywatch = {
   initialize: (serverUrl, apiKey) => {