+++ packages/react/lib/ArchivesPlayer.js
@@ -23,11 +23,11 @@ var _api = require("@skywatch/api");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-function _getRequireWildcardCache() { if (typeof WeakMap !== "function") return null; var cache = new WeakMap(); _getRequireWildcardCache = function _getRequireWildcardCache() { return cache; }; return cache; }
+function _getRequireWildcardCache(nodeInterop) { if (typeof WeakMap !== "function") return null; var cacheBabelInterop = new WeakMap(); var cacheNodeInterop = new WeakMap(); return (_getRequireWildcardCache = function _getRequireWildcardCache(nodeInterop) { return nodeInterop ? cacheNodeInterop : cacheBabelInterop; })(nodeInterop); }
 
-function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } if (obj === null || typeof obj !== "object" && typeof obj !== "function") { return { default: obj }; } var cache = _getRequireWildcardCache(); if (cache && cache.has(obj)) { return cache.get(obj); } var newObj = {}; var hasPropertyDescriptor = Object.defineProperty && Object.getOwnPropertyDescriptor; for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) { var desc = hasPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : null; if (desc && (desc.get || desc.set)) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } newObj.default = obj; if (cache) { cache.set(obj, newObj); } return newObj; }
+function _interopRequireWildcard(obj, nodeInterop) { if (!nodeInterop && obj && obj.__esModule) { return obj; } if (obj === null || typeof obj !== "object" && typeof obj !== "function") { return { default: obj }; } var cache = _getRequireWildcardCache(nodeInterop); if (cache && cache.has(obj)) { return cache.get(obj); } var newObj = {}; var hasPropertyDescriptor = Object.defineProperty && Object.getOwnPropertyDescriptor; for (var key in obj) { if (key !== "default" && Object.prototype.hasOwnProperty.call(obj, key)) { var desc = hasPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : null; if (desc && (desc.get || desc.set)) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } newObj.default = obj; if (cache) { cache.set(obj, newObj); } return newObj; }
 
-function ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }
+function ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) { symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); } keys.push.apply(keys, symbols); } return keys; }
 
 function _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(Object(source), true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(Object(source)).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }
 
@@ -40,7 +40,7 @@ const defaultPlayerOptions = {
   mobileView: false
 };
 
-const ArchivesPlayer = (_ref) => {
+const ArchivesPlayer = _ref => {
   let {
     playerOptions,
     onPlayerInit,
@@ -49,7 +49,11 @@ const ArchivesPlayer = (_ref) => {
     archiveId,
     smart_ff,
     seek,
-    style
+    style,
+    controls,
+    onTimeUpdate,
+    onEnded,
+    onReady
   } = _ref;
   const containerRef = (0, _react.useRef)(null);
   (0, _react.useEffect)(() => {
@@ -75,20 +79,28 @@ const ArchivesPlayer = (_ref) => {
     };
 
     initPlayer();
-  }, [onPlayerInit, onPlayerDispose, playerOptions]);
+  }, [onPlayerInit, onPlayerDispose, playerOptions, archiveId, seek]);
   return /*#__PURE__*/_react.default.createElement("div", {
     className: "player",
     ref: containerRef
   }, /*#__PURE__*/_react.default.createElement("video", {
     className: "video-js",
-    controls: true,
-    style: style
+    controls: controls,
+    style: style,
+    id: "archive-video",
+    onTimeUpdate: onTimeUpdate,
+    onEnded: onEnded,
+    onSeeked: () => onReady()
   }));
 };
 
 ArchivesPlayer.defaultProps = {
   seek: 0,
-  smart_ff: 0
+  smart_ff: 0,
+  controls: true,
+  onTimeUpdate: () => {},
+  onEnded: () => {},
+  onReady: () => {}
 };
 ArchivesPlayer.propTypes = {
   deviceId: _propTypes.default.string.isRequired,
@@ -98,7 +110,11 @@ ArchivesPlayer.propTypes = {
   onPlayerDispose: _propTypes.default.func.isRequired,
   seek: _propTypes.default.number,
   playerOptions: _propTypes.default.object,
-  style: _propTypes.default.object
+  style: _propTypes.default.object,
+  controls: _propTypes.default.bool,
+  onTimeUpdate: _propTypes.default.func,
+  onEnded: _propTypes.default.func,
+  onReady: _propTypes.default.func
 };
 var _default = ArchivesPlayer;
 exports.default = _default;