+++ packages/react/src/Constants.js
@@ -1,0 +1,6 @@
+// status for next archive video
+const STATUS_OK = 'ok';
+const STATUS_END = 'end';
+const STATUS_HOLE = 'hole';
+
+export {STATUS_OK, STATUS_END, STATUS_HOLE};