+++ packages/react/src/FlvPlayer.js
@@ -3,7 +3,14 @@ import PropTypes from 'prop-types';
 import flvjs from 'flv.js';
 import {Requests} from '@skywatch/api';
 
-const FlvPlayer = ({deviceId, onPlayerInit, onPlayerDispose, style}) => {
+const FlvPlayer = ({
+  deviceId,
+  onPlayerInit,
+  onPlayerDispose,
+  style,
+  controls,
+  onReady,
+}) => {
   const containerRef = useRef(null);
   useEffect(() => {
     const initPlayer = async () => {
@@ -21,7 +28,7 @@ const FlvPlayer = ({deviceId, onPlayerInit, onPlayerDispose, style}) => {
         });
         flvPlayer.attachMediaElement(videoEl);
         flvPlayer.load();
-        flvPlayer.play();
+        flvPlayer.play().then(() => onReady());
         flvPlayer.on(flvjs.Events.ERROR, (errType, errDetail) => {
           console.log(errType, errDetail);
         });
@@ -39,18 +46,23 @@ const FlvPlayer = ({deviceId, onPlayerInit, onPlayerDispose, style}) => {
 
   return (
     <div className="player" ref={containerRef}>
-      <video id="videoElement" controls muted style={style} />
+      <video id="videoElement" controls={controls} muted style={style} />
     </div>
   );
 };
 
-FlvPlayer.defaultProps = {};
+FlvPlayer.defaultProps = {
+  controls: true,
+  onReady: () => {},
+};
 
 FlvPlayer.propTypes = {
   deviceId: PropTypes.string.isRequired,
   onPlayerInit: PropTypes.func.isRequired,
   onPlayerDispose: PropTypes.func.isRequired,
   style: PropTypes.object,
+  controls: PropTypes.bool,
+  onReady: PropTypes.func,
 };
 
 export default FlvPlayer;