+++ packages/react/src/index.js
@@ -1,5 +1,7 @@
 import ArchivesPlayer from './ArchivesPlayer';
 import FlvPlayer from './FlvPlayer';
+import CameraView from './CameraView';
+import {CameraViewWebComponent} from './web-components'
 import {CoreManager, Constants} from '@skywatch/api';
 
 const {SERVER_URL, API_KEY} = Constants;
@@ -9,4 +11,10 @@ const initialize = (serverUrl, apiKey) => {
   CoreManager.set(API_KEY, apiKey);
 };
 
-export {initialize, ArchivesPlayer, FlvPlayer};
+export {
+  initialize,
+  ArchivesPlayer,
+  FlvPlayer,
+  CameraView,
+  CameraViewWebComponent
+};