+++ packages/js/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.2](https://github.com/Skywatch24/JS-Library/compare/@skywatch/js@1.2.1...@skywatch/js@1.2.2) (2021-09-17)
+
+**Note:** Version bump only for package @skywatch/js
+
+
+
+
+
+## [1.2.1](https://github.com/Skywatch24/JS-Library/compare/@skywatch/js@1.2.0...@skywatch/js@1.2.1) (2021-09-17)
+
+
+### Bug Fixes
+
+* **react:** move style directory to src/style ([2f2b6d8](https://github.com/Skywatch24/JS-Library/commit/2f2b6d8b7642c2f6f1f06f544e94e21f3acfbbf5))
+
+
+
+
+
+# [1.2.0](https://github.com/Skywatch24/JS-Library/compare/@skywatch/js@1.1.5...@skywatch/js@1.2.0) (2021-09-17)
+
+
+### Bug Fixes
+
+* resume updateCurrentTime when controls is false ([9bbba29](https://github.com/Skywatch24/JS-Library/commit/9bbba29ec6fd0f21c2ca00bd6b234b29d858f3a0))
+
+
+### Features
+
+* create CameraViewWebComponent to enable pure js usage ([9036bf9](https://github.com/Skywatch24/JS-Library/commit/9036bf9845e5780b9069ef75e02f0f9a8e65db01))
+* expose isLive function to check if the video is in live mode ([486b1e7](https://github.com/Skywatch24/JS-Library/commit/486b1e7f5d3ebc848bc059172b2b41ab4e08c013))
+
+
+
+
+
 # [1.1.0](https://github.com/Skywatch24/JS-Library/compare/@skywatch/js@1.0.4...@skywatch/js@1.1.0) (2021-05-25)